requirements.txt
aiogram
tortoise-orm
aiosqlite
python-dotenv
loguru
httpx
bs4